pyproject.toml
[project]
name = "mcptrading"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.4",
    "pandas>=2.3.0",
    "pip>=25.1.1",
    "python-dateutil>=2.9.0.post0",
    "tabulate>=0.9.0",
    "typing-extensions>=4.14.0",
    "yfinance>=0.2.63",
]
